refactor: make context_manager an internal _context_manager package
Revert the namespace-package approach in favor of the established _middleware
convention: prefix the package with an underscore and restore __init__.py files
marked 'Internal only. Do not export.' This keeps the agentic mode fully internal
while staying consistent with how the codebase marks private packages.
