+++ strands-py/src/strands/_context_manager/__init__.py
@@ -1,0 +1,4 @@
+"""Internal only. Do not export.
+
+Context management modes that operate on an agent's conversation.
+"""
