+++ strands-py/src/strands/_context_manager/modes/agentic/__init__.py
@@ -1,0 +1,4 @@
+"""Internal only. Do not export.
+
+Agentic context management: model-driven compression via injected tools.
+"""
